requirements.txt
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
tqdm>=4.66.0
python-dateutil>=2.8.2
